share GetEffectiveSeverity between AnalyzerNode and service
